adjust the size of pictures
asks the user for the new length and width of landscape and portrait photos and the dpi, then separates them into a folder for each type, and adjusts them to the needed size.  i made this to make it easier to put them into a report written in word.  no more adjusting them manually and hoping it won't wreck the formatting.